db+collector: Implement record update delta+timestamping
To relate the delta counters (for packets and bytes), to
something, we need to know when the last record were updated,
so we can determine the seconds these delta counters account for.
This way is possible to calculate bandwidth and packets per sec.

DB add timestamp "last_update" and "delta_update". Also fix type
of "delta_poll", for better precision.

Signed-off-by: Jesper Dangaard Brouer <hawk@comx.dk>