collector: Invent a general detected event transition
Simply detect if the event_state from last poll differs from the current
event_type.

This will mark the first no-signal event, but it can also be used
for drop events.  For drop events, this could be use for filtering events
for channels with excessive drops, in every poll cycle.

Signed-off-by: Jesper Dangaard Brouer <hawk@comx.dk>