Add IPv6 support
This is my first stab at adding IPv6 support to IPTV-Analyzer. I've
tried to keep down the amount of duplication for IPv4 and IPv6 and
instead use as much common code as possible but it's somewhat difficult.

I've verified it's working for both IPv4 and IPv6 streams but there
might be corner cases that aren't covered.